remove redundant check, ammend deposit function
removes redundant check on depositor account existence and ammends the
depositNonFungibleToken() function to initialized the number of token
stakers value on the token struct to 2 instead of 1. This allows the
token address to be included as the root/first staker and the depositor
or owner as the second stake in subsequent stake chains. Doing so
insures the token account will always acummulate the maximum positive
bibos. In the future the token account (or issuing contract) may be
associated with a "whitelist" which would allow for splitting of bibos
to shareholders/contributers of the token issuing contract.