Creates tokenStakers mapping
global tokenStakers maps tokenIds to an array of addresses representing
the most recent stakeChain: an ordered list of address that have
transacted with the token. By convention this list is inclusive of the
token issuing contract itself which we put at the 0 index of the list.
As a result the token contract only accumulates positive bibos upon
token transfers.
We also create a getter function to access the length of stakeChains.